DOM: merge place caret at edge functions
